ajouter les commentaires pour chaque fonction du userController

+++ src/controllers/userController.js
@@ -1,9 +1,28 @@
-const User = require('../models/userModel');
 const bcrypt = require('bcrypt');
 const jwt = require('jsonwebtoken');
-const jwtMiddleware = require('../middlewares/jwtMiddleware');
 require('dotenv').config();
 
+const User = require('../models/userModel');
+
+const jwtMiddleware = require('../middlewares/jwtMiddleware');
+
+/*
+    userRegister :
+
+    Fonction qui permet à une personne de créer un compte user
+
+    Elle prend en entrée : Un email et un mot de passe ${email} , ${password}
+
+    Les vérifications : 
+        - L'email entré correspond au format d'email
+
+
+    Reponses: 
+        201 : création du compte user. 
+            la fonction retourne l'email du user: ${email} 
+        400 : Bad request : le format de l'email n'est pas respécté
+        500 : Erreur lors du traitement de donnée
+*/
 exports.userRegister = async(req, res) =>{
     try {
         if (jwtMiddleware.verifyEmail(req.body.email)) {
@@ -17,20 +36,39 @@ exports.userRegister = async(req, res) =>{
             let user = await newUser.save();
             res.status(201).json({message: `User créer: ${user.email}`});
         } else {
-            res.status(401).json({message: 'Le format de l\'email n\'a pas été respecté'});
+            res.status(400).json({message: 'Le format de l\'email n\'a pas été respecté'});
         }
     } catch(error){
-        res.status(401).json({message: 'Requete invalide'});
+        res.status(500).json({message: 'Erreur lors du traitement des données'});
     }
 }
 
 
+/*
+    loginRegister :
+
+    Fonction qui permet à une personne de se connecter à son compte user
+
+    Elle prend en entrée : Un email et un mot de passe ${email} , ${password}
+
+    Les vérifications : 
+        - Vérifier que l'utilisateur existe dans la base de donnée
+        - Vérifier que le mot de passe fournis est le bon
+
+
+    Reponses: 
+        201 : Connection au compte user. 
+            la fonction retourne le token user: ${token} 
+        401 : Accès refusé : Mot de passe incorrect
+        404 : Utilisateur non trouvé
+        500 : Erreur lors du traitement de donnée
+*/
 exports.loginRegister = async(req, res) =>{
     try {
         const user = await User.findOne({email: req.body.email});
 
         if(!user){
-            res.status(500).json({message: "utilisateur non trouvé"});
+            res.status(404).json({message: "utilisateur non trouvé"});
             return;
         }else{
             const passwordMatch = await bcrypt.compare(req.body.password, user.password);
@@ -54,6 +92,26 @@ exports.loginRegister = async(req, res) =>{
 }
 
 
+
+/*
+    userModify :
+
+    Fonction qui permet à un user de modifier son mot de passe
+        il faut être connecté a un compte user (token user)
+
+    Elle prend en entrée : Un mot de passe ${password}
+
+    Les vérifications : 
+        - Existence du token de connection à un compte user
+        - Vérifier que l'utilisateur existe dans la base de donnée
+
+
+    Reponses: 
+        201 : Mot de passe modifié
+        403 : Accès interdit : token manquant ou expiré
+        404 : Utilisateur non trouvé
+        500 : Erreur lors du traitement de donnée
+*/
 exports.userModify = async(req, res) =>{
     try {
         let token = req.headers['authorization'];
@@ -70,16 +128,24 @@ exports.userModify = async(req, res) =>{
 
             req.user = payload;
 
-            try{
-                req.body.password = await bcrypt.hash(req.body.password, 10);
-                
-                const user = await User.findByIdAndUpdate(req.user.id, req.body, {new: true});
-                res.status(201).json({message : "votre mot de passe a bien été modifié"});
-            }
-            catch(error){
-                res.status(500);
-                    console.log(error);
-                    res.json({ message : 'Erreur serveur lors du traitement'});
+            const user = await User.findOne({_id: req.user.id});
+            //verifier l existance du user dans la bdd
+            if(user){
+                try{
+                    //hachage du password
+                    req.body.password = await bcrypt.hash(req.body.password, 10);
+                    
+                    const user = await User.findByIdAndUpdate(req.user.id, req.body, {new: true});
+                    res.status(201).json({message : "votre mot de passe a bien été modifié"});
+                }
+                catch(error){
+                    res.status(500);
+                        console.log(error);
+                        res.json({ message : 'Erreur serveur lors du traitement'});
+                }
+            }else{
+                res.status(404).json({message: "Utilisateur non trouvé"});
+                    return;
             }
         }else{
             res.status(403).json({message: "Accès interdit: token manquant"});
@@ -92,6 +158,26 @@ exports.userModify = async(req, res) =>{
 }
 
 
+
+/*
+    deleteUser :
+
+    Fonction qui permet à un user de supprié son compte
+        il faut être connecté a un compte user (token user)
+
+    Elle prend rien en entrée
+
+    Les vérifications : 
+        - Existence du token de connection à un compte user
+        - Vérifier que l'utilisateur existe dans la base de donnée
+
+
+    Reponses: 
+        201 : Compte supprimé
+        403 : Accès interdit : token manquant ou expiré
+        404 : Utilisateur non trouvé
+        500 : Erreur lors du traitement de donnée
+*/
 exports.deleteUser = async(req, res) =>{
     try {
         let token = req.headers['authorization'];
@@ -108,15 +194,22 @@ exports.deleteUser = async(req, res) =>{
 
             req.user = payload;
 
-            try{
-                const user = await User.findByIdAndDelete(req.user.id, req.body, {new: true});
-                res.status(200);
-                res.json({ message : 'Supprimé'});
-            }
-            catch(error){
-                res.status(500);
-                    console.log(error);
-                    res.json({ message : 'Erreur serveur'});
+            const user = await User.findOne({_id: req.user.id});
+
+            if(user){
+                try{
+                    const user = await User.findByIdAndDelete(req.user.id, req.body, {new: true});
+                    res.status(201);
+                    res.json({ message : 'Supprimé'});
+                }
+                catch(error){
+                    res.status(500);
+                        console.log(error);
+                        res.json({ message : 'Erreur serveur'});
+                }
+            }else{
+                res.status(404).json({message: "Utilisateur non trouvé"});
+                    return;
             }
         }else{
             res.status(403).json({message: "Accès interdit: token manquant"});
